+++ src/core/actions.ts
@@ -1,5 +1,10 @@
 import { ActionDefinitions, KeaAction, KeaReduxAction, Logic, LogicBuilder, PayloadCreatorDefinition } from '../types'
-import { getContext } from '../kea/context'
+import { getContext, getPluginContext } from '../kea/context'
+import type { ListenersPluginContext } from './listeners'
+import { isBreakpoint } from './listeners'
+
+let asyncCounter = 0
+const nextDispatchId = () => String(++asyncCounter)
 
 /** Logic builder: actions({ actionWithParams: (id) => ({ id }), actionNoParams: true }) */
 export function actions<L extends Logic = Logic>(
@@ -10,16 +15,50 @@ export function actions<L extends Logic = Logic>(
     for (const [key, payloadCreator] of Object.entries(actions)) {
       const actionCreator: KeaAction =
         typeof payloadCreator === 'function' && '_isKeaAction' in payloadCreator
-          ? payloadCreator
+          ? (payloadCreator as KeaAction)
           : createActionCreator(createActionType(key, logic.pathString), payloadCreator ?? true)
       const type = actionCreator.toString()
+
+      const {
+        options: { disableAsyncActions },
+      } = getContext()
 
       logic.actionCreators[key] = actionCreator
       logic.actions[key] = (...inp: any[]) => {
         const builtAction = actionCreator(...inp)
-        getContext().store.dispatch(builtAction)
+        getContext().store.dispatch(
+          disableAsyncActions ? builtAction : { ...builtAction, dispatchId: nextDispatchId() },
+        )
       }
       logic.actions[key].toString = () => type
+      if (!disableAsyncActions) {
+        logic.asyncActions[key] = async (...inp: any[]) => {
+          const builtAction = actionCreator(...inp)
+          let dispatchId = nextDispatchId()
+          getContext().store.dispatch({ ...builtAction, dispatchId })
+          const { pendingDispatches } = getPluginContext<ListenersPluginContext>('listeners')
+          while (true) {
+            const promises = pendingDispatches.get(dispatchId)
+            if (!promises) {
+              return
+            }
+            try {
+              const responses = await Promise.all(promises)
+              return responses[0]
+            } catch (e: any) {
+              if (isBreakpoint(e)) {
+                if ('__keaDispatchId' in e) {
+                  dispatchId = e.__keaDispatchId
+                  // loop again
+                }
+              } else {
+                throw e
+              }
+            }
+          }
+        }
+        logic.asyncActions[key].toString = () => type
+      }
       logic.actionKeys[type] = key
       logic.actionTypes[key] = type
     }